requirements.txt
pygame>=2.5.0
clipspy>=1.0.0
